+++ packages/purpose-readmodel-writer/src/purposeConsumerServiceV2.ts
